TarotApp_server/requirements.txt
Flask==3.1.1
Flask-RESTful==0.3.10
llama-cpp-python==0.3.9
Werkzeug==3.1.3
wheel==0.45.1